feat: reintroduce bip85 seed unwrap for the sake of temp backawrds compatibility
