+++ ext/package.json
@@ -37,6 +37,7 @@
     "bip21": "3.0.0",
     "bip32": "4.0.0",
     "bip39": "3.1.0",
+    "bip85": "0.0.3",
     "bitcoinjs-lib": "6.1.6",
     "bitcoinjs-message": "2.2.0",
     "bolt11": "^1.4.1",
