+++ mobile/package.json
@@ -22,7 +22,7 @@
     "web": "expo start --web",
     "test": "jest --watchAll",
     "e2e": "maestro test .maestro/*",
-    "unit": "vitest run ../shared/tests/unit-vi/ --silent",
+    "unit": "vitest run ../shared/tests/unit-vi/ ./src/tests/unit-vi/ --silent",
     "integration": "vitest run ../shared/tests/integration-vi/ --silent",
     "lint": "prettier '**/*.{js,jsx,ts,tsx,json,css,scss,md,yml}' '../shared/**/*.{js,jsx,ts,tsx,json,css,scss,md,yml}' --config ./.prettierrc --check && tsc --noEmit && expo lint",
     "lint:write": "prettier --write '**/*.{js,jsx,ts,tsx,json,css,scss,md,yml}' '../shared/**/*.{js,jsx,ts,tsx,json,css,scss,md,yml}' --config ./.prettierrc; expo lint --fix",
@@ -56,6 +56,7 @@
     "bip21": "3.0.0",
     "bip32": "4.0.0",
     "bip39": "3.1.0",
+    "bip85": "0.0.3",
     "bitcoinjs-lib": "^6.1.7",
     "bitcoinjs-message": "^2.2.0",
     "bolt11": "^1.4.1",
